refactored helper functions: moved helpers formatDate.js and getAuthFormContent to shared/functions
