+++ src/components/Authorization/RegisterForm/RegisterForm.jsx
@@ -52,7 +52,11 @@ export const RegisterForm = () => {
 			phone,
 		};
 		console.log(res);
-		dispatch(reg(res)).unwrap().then(()=> toast.success('Welcome!')).catch((e) => {
+		dispatch(reg(res)).unwrap()
+			.then((res) => {
+			const {name}= res.user
+			return toast.success(`Welcome, ${name} !`) })
+			.catch((e) => {
 			const errorMessage = e.status === 409 ? 'This email is already in use' : 'Oops, something went wrong... Try again, please'
 			return toast.error(errorMessage)
 		});
