+++ src/shared/functions/getAuthFormContent.js
